Fix forward reference Python 3.8 issue
