+++ tests/test_stacktrace.py
@@ -1,9 +1,54 @@
+from __future__ import annotations
 import json
+from typing import Optional, List
 
 from implicitdict import ImplicitDict
 
 import pytest
-from .test_types import MassiveNestingData
+
+
+# This object must be defined with future annotations as Python 3.8 will not resolve string-based forward references correctly
+class MassiveNestingData(ImplicitDict):
+    children: Optional[List[MassiveNestingData]]
+    foo: str
+    bar: int = 0
+
+    @staticmethod
+    def example_value():
+        return ImplicitDict.parse(
+            {
+                "foo": "1a",
+                "children": [
+                    {
+                        "foo": "1a 2a"
+                    },
+                    {
+                        "foo": "1a 2b",
+                        "children": [
+                            {
+                                "foo": "1a 2b 3",
+                                "children": [
+                                    {
+                                        "foo": "1a 2b 3 4a",
+                                        "bar": 123
+                                    },
+                                    {
+                                        "foo": "1a 2b 3 4b",
+                                        "bar": 456
+                                    },
+                                    {
+                                        "foo": "1a 2b 3 4c",
+                                        "bar": 789,
+                                        "children": []
+                                    }
+                                ]
+                            }
+                        ]
+                    }
+                ]
+            },
+            MassiveNestingData
+        )
 
 
 def _get_correct_value() -> MassiveNestingData:
